added exceptions

+++ src/controller/UserController.java
@@ -1,8 +1,13 @@
 package controller;
 
+import exceptions.EntityNotFound;
+import model.Customer;
+import model.Delivery_Person;
+import model.Employee;
 import service.UserService;
 
 import java.util.List;
+import java.util.Objects;
 
 public class UserController {
     private UserService userService;
@@ -39,6 +44,18 @@ public class UserController {
      * @param customerId the ID of the customer to delete
      */
     public void deleteCustomer(Integer customerId) {
+        List<Customer> customers = userService.getCustomers();
+        boolean existsCustomer = false;
+
+        for (Customer customer: customers){
+            if (customer.getCustomerID() == customerId){
+                existsCustomer = true;
+                break;
+            }
+        }
+
+        if(!existsCustomer) throw new EntityNotFound("No customer found with ID " + customerId);
+
         userService.deleteCustomer(customerId);
         System.out.println("Deleted customer with id " + customerId + " successfully");
     }
@@ -67,6 +84,18 @@ public class UserController {
      * @param employeeId the ID of the employee to delete
      */
     public void deleteEmployee(Integer employeeId) {
+        List<Employee> employees = userService.getEmployees();
+        boolean existsEmployee = false;
+
+        for (Employee employee: employees){
+            if (employee.getEmployeeID() == employeeId){
+                existsEmployee = true;
+                break;
+            }
+        }
+
+        if(!existsEmployee) throw new EntityNotFound("No employee found with ID " + employeeId);
+
         userService.unenrollEmployee(employeeId);
         System.out.println("Employee with id " + employeeId + " unenrolled successfully");
     }
@@ -86,6 +115,18 @@ public class UserController {
      * @param deliveryPersonId the ID of the delivery person to delete
      */
     public void deleteDeliveryPerson(Integer deliveryPersonId) {
+        List<Delivery_Person> deliveryPeople = userService.getDeliveryPerson();
+        boolean existsDeliveryPerson = false;
+
+        for (Delivery_Person deliveryPerson: deliveryPeople){
+            if (Objects.equals(deliveryPerson.getId(), deliveryPersonId)){
+                existsDeliveryPerson = true;
+                break;
+            }
+        }
+
+        if (!existsDeliveryPerson) throw new EntityNotFound("No delivery person found with ID " + deliveryPersonId);
+
         userService.unenrollDeliveryPerson(deliveryPersonId);
         System.out.println("Delivery person with id " + deliveryPersonId + " unenrolled successfully");
     }
